Change the random content path each time new files are shared.
This protects against a potential attack where the attacker controls another app on the same device as OnionShare, and has also received the URL of one of the user's previous shares. The attacker would then be able to reuse the random content path from the known URL to download future shares via the other app.